feat: UserService 추상화 및 App, Tx UserService 구현
